+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/fx/Features.java
@@ -2,6 +2,7 @@ package de.mephisto.vpin.commons.fx;
 
 public class Features {
   public static boolean MANIA_ENABLED = true;
+  public static boolean MANIA_SOCIAL_ENABLED = true;
   public static boolean ISCORED_ENABLED = true;
   public static boolean NOTIFICATIONS_ENABLED = true;
   public static boolean VPX_MONITORING = true;