+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/highscores/cards/CardService.java
@@ -30,7 +30,9 @@ import org.springframework.stereotype.Service;
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.util.Arrays;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
@@ -53,6 +55,8 @@ public class CardService implements InitializingBean, HighscoreChangeListener, P
   private CardTemplatesService cardTemplatesService;
 
   private CardSettings cardSettings;
+
+  private Map<Integer, ScoreSummary> scoreCache = new LinkedHashMap<>();
 
   public File generateTableCardFile(Game game) {
     generateCard(game);
@@ -110,8 +114,7 @@ public class CardService implements InitializingBean, HighscoreChangeListener, P
    */
   public boolean generateCard(Game game, boolean generateSampleCard, CardTemplate template) {
     try {
-      long serverId = preferencesService.getPreferenceValueLong(PreferenceNames.DISCORD_GUILD_ID, -1);
-      ScoreSummary summary = highscoreService.getScoreSummary(serverId, game);
+      ScoreSummary summary = getScoreSummary(game, template, generateSampleCard);
       Platform.runLater(() -> {
         Thread.currentThread().setName("FX Card Generator Thread for " + game.getGameDisplayName());
         doGenerateCard(game, summary, generateSampleCard, template);
@@ -127,6 +130,29 @@ public class CardService implements InitializingBean, HighscoreChangeListener, P
       LOG.error("Failed to generate image: " + e.getMessage(), e);
       return false;
     }
+  }
+
+  @NonNull
+  private ScoreSummary getScoreSummary(Game game, CardTemplate template, boolean generateSampleCard) {
+    long serverId = preferencesService.getPreferenceValueLong(PreferenceNames.DISCORD_GUILD_ID, -1);
+    ScoreSummary summary = null;
+    if (template.isRenderFriends()) {
+      //add simply caching until a real card is generated, should be sufficient while editing
+      if (generateSampleCard) {
+        if (!scoreCache.containsKey(game.getId())) {
+          scoreCache.put(game.getId(), highscoreService.getMergedScoreSummary(serverId, game));
+        }
+
+        return scoreCache.get(game.getId());
+      }
+
+      scoreCache.clear();
+      summary = highscoreService.getMergedScoreSummary(serverId, game);
+    }
+    else {
+      summary = highscoreService.getScoreSummary(serverId, game);
+    }
+    return summary;
   }
 
   private boolean doGenerateCard(Game game, ScoreSummary summary, boolean generateSampleCard, CardTemplate template) {