+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/util/AvatarFactory.java
@@ -15,25 +15,30 @@ import java.io.InputStream;
 
 public class AvatarFactory {
 
-  public static Control create(Image image) {
-    Tile avatar = TileBuilder.create()
-      .skinType(Tile.SkinType.IMAGE)
-      .maxSize(UIDefaults.DEFAULT_AVATARSIZE, UIDefaults.DEFAULT_AVATARSIZE)
-      .backgroundColor(Color.TRANSPARENT)
-      .image(image)
-      .imageMask(Tile.ImageMask.ROUND)
-      .textSize(Tile.TextSize.BIGGER)
-      .textAlignment(TextAlignment.CENTER)
-      .build();
-    return avatar;
-  }
-
   public static ImageView create(InputStream in) {
     if (in == null) {
       in = Studio.class.getResourceAsStream("avatar-blank.png");
     }
 
     Image image = new Image(in);
+    ImageView imageView = new ImageView(image);
+    imageView.setFitWidth(UIDefaults.DEFAULT_AVATARSIZE);
+    imageView.setPreserveRatio(true);
+    imageView.setSmooth(true);
+
+    Rectangle clip = new Rectangle();
+    clip.setWidth(UIDefaults.DEFAULT_AVATARSIZE);
+    clip.setHeight(UIDefaults.DEFAULT_AVATARSIZE);
+
+    clip.setArcHeight(UIDefaults.DEFAULT_AVATARSIZE);
+    clip.setArcWidth(UIDefaults.DEFAULT_AVATARSIZE);
+    clip.setStroke(Color.WHITE);
+    clip.setStrokeWidth(0);
+    imageView.setClip(clip);
+    return imageView;
+  }
+
+  public static ImageView createAvatarImageView(Image image) {
     ImageView imageView = new ImageView(image);
     imageView.setFitWidth(UIDefaults.DEFAULT_AVATARSIZE);
     imageView.setPreserveRatio(true);